+++ CHANGELOG.md
@@ -15,6 +15,12 @@
 
 * Fixed 'flickering' stories when an uploaded image replaced a placeholder
 
+## [0.2.1] - 2025-03-30
+
+### Added
+
+* Added interface for listening to story reader events
+
 ## [0.2.0] - 2025-04-25
 
 ### Added